% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_normal_n_in_interval}
\alias{generate_normal_n_in_interval}
\title{Generate Normally Distributed Data Within a Specified Interval}
\usage{
generate_normal_n_in_interval(n, mean, sd, min_val, max_val)
}
\arguments{
\item{n}{The number of data points to generate. Must be a positive integer.}

\item{mean}{The mean of the normal distribution.}

\item{sd}{The standard deviation of the normal distribution. Must be a positive number.}

\item{min_val}{The minimum value of the interval. Must be less than `max_val`.}

\item{max_val}{The maximum value of the interval. Must be greater than `min_val`.}
}
\value{
A numeric vector of length `n` containing the generated data points that fall within the specified interval.
}
\description{
This function generates a specified number of data points from a normal distribution
that fall within a given interval. It ensures that all generated data points are
within the specified minimum and maximum values.
}
\examples{
generated_data <- generate_normal_n_in_interval(n = 100, mean = 0, sd = 1, min_val = -2, max_val = 2)
hist(generated_data)

}
